fix the constant in BlocksPool and changed some comments

+++ core/src/test/java/com/oath/oak/NativeAllocator/OakNativeMemoryAllocatorTest.java
@@ -87,7 +87,7 @@ public class OakNativeMemoryAllocatorTest {
     @Test
     public void checkCapacity() {
 
-        int blockSize = (int) BlocksPool.getInstance().blockSize();
+        int blockSize = BlocksPool.getInstance().blockSize();
         int capacity = blockSize * 3;
         OakNativeMemoryAllocator ma = new OakNativeMemoryAllocator(capacity);
